feat: 更新角色属性系统并补充 CharacterDataSO 全属性字段

+++ Assets/Scripts/DamageInfo.cs
@@ -2,11 +2,11 @@ using UnityEngine;
 
 public struct DamageInfo
 {
-    public int damage;
+    public float damage;
     public Vector2 position; 
     public bool isCritical;
 
-    public DamageInfo(int damage, Vector2 position, bool isCritical)
+    public DamageInfo(float damage, Vector2 position, bool isCritical)
     {
         this.damage = damage;
         this.position = position;
